rfc: separate concerns in settings
